Modify AccountController
