+++ BusinessLogic/Repositories/ShopRepository.cs
@@ -7,16 +7,10 @@ namespace ClothingShop.BusinessLogic.Repositories
     public class ShopRepository : IShopRepository
     {
         private readonly ShopContext _db;
-        
 
         public ShopRepository(ShopContext db)
         {
             _db = db;
-        }
-        public string getUser(string email)
-        {
-            
-            return "";
         }
     }
 }
