+++ ClothingShop/Controllers/AccountController.cs
@@ -1,0 +1,140 @@
+﻿using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
+using Microsoft.AspNetCore.Mvc;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+using ClothingShop.BusinessLogic.Repositories;
+using ClothingShop.Entity.Models;
+using System.Security.Cryptography;
+using System.Text;
+using ClothingShop.Entity.Data;
+using ClothingShop.BusinessLogic.Repositories.Interfaces;
+using ClothingShop.Entity.Entities;
+
+namespace ClothingShop.Controllers
+{
+    public class AccountController : Controller
+    {
+        private readonly SignInManager<Users> signInManager;
+        private readonly UserManager<Users> userManager;
+
+        public AccountController(SignInManager<Users> signInManager,
+                                 UserManager<Users> userManager)
+        {
+            this.signInManager = signInManager;
+            this.userManager = userManager;
+        }
+
+        public IActionResult AccessDenied()
+        {
+            return View();
+        }
+
+        public IActionResult Index()
+        {
+            return RedirectToAction("Login");
+        }
+
+        //Login
+        [HttpGet]
+        public IActionResult Login(string returnUrl)
+        {
+            ViewData["ReturnUrl"] = returnUrl;
+            return View();
+        }
+
+        [HttpPost]
+        [AllowAnonymous]
+        [ValidateAntiForgeryToken]
+        public async Task<IActionResult> Login(LoginModel model, string returnUrl = null)
+        {
+            ViewData["ReturnUrl"] = returnUrl;
+            if (ModelState.IsValid)
+            {
+                var result = await signInManager.PasswordSignInAsync(model.UserName, model.Password, model.RememberMe, lockoutOnFailure: false).ConfigureAwait(false);
+                if (result.Succeeded)
+                {
+                    return RedirectToLocal(returnUrl);
+                }
+                else
+                {
+                    ModelState.AddModelError(string.Empty, "Invalid login attempt.");
+                    return View(model);
+                }
+            }
+            return View(model);
+        }
+
+        //Register
+        [HttpGet]
+        public IActionResult CreateAccount()
+        {
+            return View();
+        }
+
+        [HttpPost]
+        [ValidateAntiForgeryToken]
+        public async Task<IActionResult> CreateAccount(LoginModel model)
+        {
+            if (ModelState.IsValid)
+            {
+                var user = new Users
+                {
+                    UserName = model.UserName,
+                    Email = model.UserName
+                };
+                var hash_pass = GetMD5(model.Password);
+                var result = await userManager.CreateAsync(user, hash_pass);
+                if (result.Succeeded)
+                {
+                    await signInManager.SignInAsync(user, isPersistent: false);
+                    return RedirectToAction("index", "home");
+                }
+                foreach (var error in result.Errors)
+                {
+                    ModelState.AddModelError("", error.Description);
+                }
+            }
+            return View(model);
+        }
+
+        //SignOff
+        [HttpPost]
+        [ValidateAntiForgeryToken]
+        public async Task<IActionResult> SignOff()
+        {
+            await signInManager.SignOutAsync().ConfigureAwait(false);
+            return RedirectToAction("Login");
+        }
+
+        private IActionResult RedirectToLocal(string returnUrl)
+        {
+            if (Url.IsLocalUrl(returnUrl))
+            {
+                return Redirect(returnUrl);
+            }
+            else
+            {
+                return RedirectToAction(nameof(HomeController.Index), "Home");
+            }
+        }
+
+        //!!! MOVE TO REPOSITORY
+        //Create a string MD5
+        public static string GetMD5(string str)
+        {
+            MD5 md5 = new MD5CryptoServiceProvider();
+            byte[] fromData = Encoding.UTF8.GetBytes(str);
+            byte[] targetData = md5.ComputeHash(fromData);
+            string byte2String = null;
+
+            for (int i = 0; i < targetData.Length; i++)
+            {
+                byte2String += targetData[i].ToString("x2");
+            }
+            return byte2String;
+        }
+    }
+}
